Move FTS filerep specific functions to separate file.
Also, fixes bunch of warnings introduced due to fts walrep work.
