+++ src/backend/fts/Makefile
@@ -11,6 +11,10 @@ top_builddir = ../../..
 include $(top_builddir)/src/Makefile.global
 override CPPFLAGS := -I$(top_srcdir)/src/backend/gp_libpq_fe $(CPPFLAGS)
 
-OBJS = fts.o ftsprobe.o ftsfilerep.o ftsprobehandler.o
-       
+OBJS = fts.o ftsprobe.o
+ifeq ($(enable_segwalrep), yes)
+OBJS += ftsprobehandler.o
+else
+OBJS += ftsfilerep.o ftsprobefilerep.o
+endif
 include $(top_srcdir)/src/backend/common.mk
