+++ src/backend/fts/ftsfilerep.c
@@ -16,15 +16,21 @@
  */
 #include "postgres.h"
 
-#include "gp-libpq-fe.h"
-#include "gp-libpq-int.h"
-#include "cdb/cdbfts.h"
-#include "executor/spi.h"
-#include "postmaster/fts.h"
-#include "postmaster/primary_mirror_mode.h"
+#include "access/heapam.h"
+#include "access/genam.h"
+#include "catalog/gp_configuration_history.h"
+#include "catalog/gp_segment_config.h"
 
 #include "cdb/ml_ipc.h" /* gettime_elapsed_ms */
-
+#include "executor/spi.h"
+#include "gp-libpq-fe.h"
+#include "gp-libpq-int.h"
+#include "postmaster/fts.h"
+#include "postmaster/primary_mirror_mode.h"
+#include "utils/fmgroids.h"
+#include "utils/snapmgr.h"
+#include "utils/syscache.h"
+#include "cdb/cdbfts.h"
 
 /*
  * CONSTANTS
@@ -269,6 +275,143 @@ FtsResolveStateFilerep(FtsSegmentPairState *pairState)
 	}
 }
 
+/*
+ * Marks the given db as in-sync in the segment configuration.
+ */
+static void
+FtsMarkSegmentsInSync(CdbComponentDatabaseInfo *primary, CdbComponentDatabaseInfo *mirror)
+{
+	if (!FTS_STATUS_ISALIVE(primary->dbid, ftsProbeInfo->fts_status) ||
+	    !FTS_STATUS_ISALIVE(mirror->dbid, ftsProbeInfo->fts_status) ||
+	    !FTS_STATUS_ISPRIMARY(primary->dbid, ftsProbeInfo->fts_status) ||
+ 	    FTS_STATUS_ISPRIMARY(mirror->dbid, ftsProbeInfo->fts_status) ||
+	    FTS_STATUS_IS_SYNCED(primary->dbid, ftsProbeInfo->fts_status) ||
+	    FTS_STATUS_IS_SYNCED(mirror->dbid, ftsProbeInfo->fts_status) ||
+	    FTS_STATUS_IS_CHANGELOGGING(primary->dbid, ftsProbeInfo->fts_status) ||
+	    FTS_STATUS_IS_CHANGELOGGING(mirror->dbid, ftsProbeInfo->fts_status))
+	{
+		FtsRequestPostmasterShutdown(primary, mirror);
+	}
+
+	if (ftsProbeInfo->fts_pauseProbes)
+	{
+		return;
+	}
+
+	uint8	segStatus=0;
+	Relation configrel;
+	Relation histrel;
+	ScanKeyData scankey;
+	SysScanDesc sscan;
+	HeapTuple configtuple;
+	HeapTuple newtuple;
+	HeapTuple histtuple;
+	Datum configvals[Natts_gp_segment_configuration];
+	bool confignulls[Natts_gp_segment_configuration] = { false };
+	bool repls[Natts_gp_segment_configuration] = { false };
+	Datum histvals[Natts_gp_configuration_history];
+	bool histnulls[Natts_gp_configuration_history] = { false };
+	char *desc = "FTS: changed segment to insync from resync.";
+	/*
+	 * Commit/abort transaction below will destroy
+	 * CurrentResourceOwner.  We need it for catalog reads.
+	 */
+	ResourceOwner save = CurrentResourceOwner;
+	StartTransactionCommand();
+	GetTransactionSnapshot();
+
+	/* update primary */
+	segStatus = ftsProbeInfo->fts_status[primary->dbid];
+	segStatus |= FTS_STATUS_SYNCHRONIZED;
+	ftsProbeInfo->fts_status[primary->dbid] = segStatus;
+
+	/* update mirror */
+	segStatus = ftsProbeInfo->fts_status[mirror->dbid];
+	segStatus |= FTS_STATUS_SYNCHRONIZED;
+	ftsProbeInfo->fts_status[mirror->dbid] = segStatus;
+
+	histrel = heap_open(GpConfigHistoryRelationId,
+						RowExclusiveLock);
+	configrel = heap_open(GpSegmentConfigRelationId,
+						  RowExclusiveLock);
+
+	/* update gp_segment_configuration to insync */
+	ScanKeyInit(&scankey,
+				Anum_gp_segment_configuration_dbid,
+				BTEqualStrategyNumber, F_INT2EQ,
+				Int16GetDatum(primary->dbid));
+	sscan = systable_beginscan(configrel, GpSegmentConfigDbidIndexId,
+							   true, SnapshotNow, 1, &scankey);
+	configtuple = systable_getnext(sscan);
+	if (!HeapTupleIsValid(configtuple))
+	{
+		elog(ERROR,"FTS cannot find dbid (%d, %d) in %s", primary->dbid,
+			 mirror->dbid, RelationGetRelationName(configrel));
+	}
+	configvals[Anum_gp_segment_configuration_mode-1] = CharGetDatum(GP_SEGMENT_CONFIGURATION_MODE_INSYNC);
+	repls[Anum_gp_segment_configuration_mode-1] = true;
+	newtuple = heap_modify_tuple(configtuple, RelationGetDescr(configrel),
+								 configvals, confignulls, repls);
+	simple_heap_update(configrel, &configtuple->t_self, newtuple);
+	CatalogUpdateIndexes(configrel, newtuple);
+
+	systable_endscan(sscan);
+
+	ScanKeyInit(&scankey,
+				Anum_gp_segment_configuration_dbid,
+				BTEqualStrategyNumber, F_INT2EQ,
+				Int16GetDatum(mirror->dbid));
+	sscan = systable_beginscan(configrel, GpSegmentConfigDbidIndexId,
+							   true, SnapshotNow, 1, &scankey);
+	configtuple = systable_getnext(sscan);
+	if (!HeapTupleIsValid(configtuple))
+	{
+		elog(ERROR,"FTS cannot find dbid (%d, %d) in %s", primary->dbid,
+			 mirror->dbid, RelationGetRelationName(configrel));
+	}
+	newtuple = heap_modify_tuple(configtuple, RelationGetDescr(configrel),
+								 configvals, confignulls, repls);
+	simple_heap_update(configrel, &configtuple->t_self, newtuple);
+	CatalogUpdateIndexes(configrel, newtuple);
+
+	systable_endscan(sscan);
+
+	/* update configuration history */
+	histvals[Anum_gp_configuration_history_time-1] =
+			TimestampTzGetDatum(GetCurrentTimestamp());
+	histvals[Anum_gp_configuration_history_dbid-1] =
+			Int16GetDatum(primary->dbid);
+	histvals[Anum_gp_configuration_history_desc-1] =
+				CStringGetTextDatum(desc);
+	histtuple = heap_form_tuple(RelationGetDescr(histrel), histvals, histnulls);
+	simple_heap_insert(histrel, histtuple);
+	CatalogUpdateIndexes(histrel, histtuple);
+
+	histvals[Anum_gp_configuration_history_dbid-1] =
+			Int16GetDatum(mirror->dbid);
+	histtuple = heap_form_tuple(RelationGetDescr(histrel), histvals, histnulls);
+	simple_heap_insert(histrel, histtuple);
+	CatalogUpdateIndexes(histrel, histtuple);
+	ereport(LOG,
+			(errmsg("FTS: resynchronization of mirror (dbid=%d, content=%d) on %s:%d has completed.",
+					mirror->dbid, mirror->segindex, mirror->address, mirror->port ),
+			 errSendAlert(true)));
+
+	heap_close(histrel, RowExclusiveLock);
+	heap_close(configrel, RowExclusiveLock);
+	/*
+	 * Do not block shutdown.  We will always get a change to update
+	 * gp_segment_configuration in subsequent probes upon database
+	 * restart.
+	 */
+	if (IsFtsShudownRequested())
+	{
+		elog(LOG, "Shutdown in progress, ignoring FTS prober updates.");
+		return;
+	}
+	CommitTransactionCommand();
+	CurrentResourceOwner = save;
+}
 
 /*
  * pre-process probe results to take into account some special
@@ -283,11 +426,11 @@ FtsPreprocessProbeResultsFilerep(CdbComponentDatabases *dbs, uint8 *probe_result
 {
 	int i = 0;
 	cdb_component_dbs = dbs;
-	Assert(cdb_component_dbs != NULL);
+	Assert(dbs != NULL);
 
-	for (i=0; i < cdb_component_dbs->total_segment_dbs; i++)
+	for (i=0; i < dbs->total_segment_dbs; i++)
 	{
-		CdbComponentDatabaseInfo *segInfo = &cdb_component_dbs->segment_db_info[i];
+		CdbComponentDatabaseInfo *segInfo = &dbs->segment_db_info[i];
 		CdbComponentDatabaseInfo *primary = NULL, *mirror = NULL;
 
 		if (!SEGMENT_IS_ACTIVE_PRIMARY(segInfo))
@@ -568,6 +711,7 @@ getHostsByDbid(int dbid, char **hostname_p, int *host_port_p, int *host_filerep_
 	Assert(peer_name_p != NULL);
 	Assert(peer_pm_port_p != NULL);
 	Assert(peer_filerep_port_p != NULL);
+	Assert(cdb_component_dbs != NULL);
 
 	found = false;
 
@@ -622,5 +766,379 @@ getHostsByDbid(int dbid, char **hostname_p, int *host_port_p, int *host_filerep_
 	}
 }
 
+#ifndef USE_SEGWALREP
+static uint32 getTransition(bool isPrimaryAlive, bool isMirrorAlive);
+
+static void
+buildSegmentStateChange
+	(
+	CdbComponentDatabaseInfo *segInfo,
+	FtsSegmentStatusChange *change,
+	uint8 statusNew
+	)
+	;
+
+static uint32 transition
+	(
+	uint32 stateOld,
+	uint32 trans,
+	CdbComponentDatabaseInfo *primary,
+	CdbComponentDatabaseInfo *mirror,
+	FtsSegmentStatusChange *changesPrimary,
+	FtsSegmentStatusChange *changesMirror
+	)
+	;
+
+static void updateConfiguration(FtsSegmentStatusChange *changes, int changeEntries);
+
+/*
+ * Build a set of changes, based on our current state, and the probe results.
+ */
+bool probePublishUpdate(CdbComponentDatabases *dbs, uint8 *probe_results)
+{
+	bool update_found = false;
+	int i;
+
+	Assert(dbs != NULL);
+	cdb_component_dbs = dbs;
+
+	/* preprocess probe results to decide what is the current segment state */
+	FtsPreprocessProbeResultsFilerep(dbs, probe_results);
+
+	for (i = 0; i < dbs->total_segment_dbs; i++)
+	{
+		CdbComponentDatabaseInfo *segInfo = &dbs->segment_db_info[i];
+
+		/* if we've gotten a pause or shutdown request, we ignore our probe results. */
+		if (!FtsIsActive())
+		{
+			return false;
+		}
+
+		/* we check segments in pairs of primary-mirror */
+		if (!SEGMENT_IS_ACTIVE_PRIMARY(segInfo))
+		{
+			continue;
+		}
+
+		CdbComponentDatabaseInfo *primary = segInfo;
+		CdbComponentDatabaseInfo *mirror = FtsGetPeerSegment(segInfo->segindex, segInfo->dbid);
+
+		Assert(mirror != NULL);
+
+		/* changes required for primary and mirror */
+		FtsSegmentStatusChange changes[2];
+
+		uint32 stateOld = 0;
+		uint32 stateNew = 0;
+
+		bool isPrimaryAlive = PROBE_IS_ALIVE(primary);
+		bool isMirrorAlive = PROBE_IS_ALIVE(mirror);
+
+		/* get transition type */
+		uint32 trans = getTransition(isPrimaryAlive, isMirrorAlive);
+
+		if (gp_log_fts > GPVARS_VERBOSITY_VERBOSE)
+		{
+			elog(LOG, "FTS: primary found %s, mirror found %s, transition %d.",
+				 (isPrimaryAlive ? "alive" : "dead"), (isMirrorAlive ? "alive" : "dead"), trans);
+		}
+
+		if (trans == TRANS_D_D)
+		{
+			elog(LOG, "FTS: detected double failure for content=%d, primary (dbid=%d), mirror (dbid=%d).",
+			     primary->segindex, primary->dbid, mirror->dbid);
+		}
+
+			/* get current state */
+		stateOld = FtsGetPairStateFilerep(primary, mirror);
+
+		/* get new state */
+		stateNew = transition(stateOld, trans, primary, mirror, &changes[0], &changes[1]);
+
+		/* check if transition is required */
+		if (stateNew != stateOld)
+		{
+			update_found = true;
+			updateConfiguration(changes, ARRAY_SIZE(changes));
+		}
+	}
+
+	if (gp_log_fts >= GPVARS_VERBOSITY_VERBOSE)
+	{
+		elog(LOG, "FTS: probe result processing is complete.");
+	}
+
+	return update_found;
+}
+
+/*
+ * Build struct with segment changes
+ */
+static void
+buildSegmentStateChange(CdbComponentDatabaseInfo *segInfo, FtsSegmentStatusChange *change, uint8 statusNew)
+{
+	change->dbid = segInfo->dbid;
+	change->segindex = segInfo->segindex;
+	change->oldStatus = ftsProbeInfo->fts_status[segInfo->dbid];
+	change->newStatus = statusNew;
+}
+
+/*
+ * get transition type - derived from probed primary/mirror state
+ */
+static uint32
+getTransition(bool isPrimaryAlive, bool isMirrorAlive)
+{
+	uint32 state = (isPrimaryAlive ? 2 : 0) + (isMirrorAlive ? 1 : 0);
+
+	switch (state)
+	{
+		case (0):
+			/* primary and mirror dead */
+			return TRANS_D_D;
+		case (1):
+			/* primary dead, mirror alive */
+			return TRANS_D_U;
+		case (2):
+			/* primary alive, mirror dead */
+			return TRANS_U_D;
+		case (3):
+			/* primary and mirror alive */
+			return TRANS_U_U;
+		default:
+			Assert(!"Invalid transition for FTS state machine");
+			return 0;
+	}
+}
+
+/*
+ * find new state for primary and mirror
+ */
+static uint32
+transition
+	(
+	uint32 stateOld,
+	uint32 trans,
+	CdbComponentDatabaseInfo *primary,
+    CdbComponentDatabaseInfo *mirror,
+    FtsSegmentStatusChange *changesPrimary,
+    FtsSegmentStatusChange *changesMirror
+    )
+{
+	Assert(IS_VALID_TRANSITION(trans));
+
+	/* reset changes */
+	memset(changesPrimary, 0, sizeof(*changesPrimary));
+	memset(changesMirror, 0, sizeof(*changesMirror));
+
+	uint32 stateNew = stateOld;
+
+	/* in case of a double failure we don't do anything */
+	if (trans == TRANS_D_D)
+	{
+		return stateOld;
+	}
+
+	/* get new state for primary and mirror */
+	stateNew = FtsTransitionFilerep(stateOld, trans);
+
+	/* check if transition is required */
+	if (stateNew != stateOld)
+	{
+		FtsSegmentPairState pairState;
+		memset(&pairState, 0, sizeof(pairState));
+		pairState.primary = primary;
+		pairState.mirror = mirror;
+		pairState.stateNew = stateNew;
+		pairState.statePrimary = 0;
+		pairState.stateMirror = 0;
+
+		if (gp_log_fts >= GPVARS_VERBOSITY_DEBUG)
+		{
+			elog(LOG, "FTS: state machine transition from %d to %d.", stateOld, stateNew);
+		}
+
+		FtsResolveStateFilerep(&pairState);
+
+		buildSegmentStateChange(primary, changesPrimary, pairState.statePrimary);
+		buildSegmentStateChange(mirror, changesMirror, pairState.stateMirror);
+
+		FtsDumpChanges(changesPrimary, 1);
+		FtsDumpChanges(changesMirror, 1);
+	}
+
+	return stateNew;
+}
+
+/*
+ * update segment configuration in catalog and shared memory
+ */
+static bool
+probeUpdateConfig(FtsSegmentStatusChange *changes, int changeCount)
+{
+	Relation configrel;
+	Relation histrel;
+	SysScanDesc sscan;
+	ScanKeyData scankey;
+	HeapTuple configtuple;
+	HeapTuple newtuple;
+	HeapTuple histtuple;
+	Datum configvals[Natts_gp_segment_configuration];
+	bool confignulls[Natts_gp_segment_configuration] = { false };
+	bool repls[Natts_gp_segment_configuration] = { false };
+	Datum histvals[Natts_gp_configuration_history];
+	bool histnulls[Natts_gp_configuration_history] = { false };
+	bool valid;
+	bool primary;
+	bool changelogging;
+	int i;
+	char desc[SQL_CMD_BUF_SIZE];
+
+	/*
+	 * Commit/abort transaction below will destroy
+	 * CurrentResourceOwner.  We need it for catalog reads.
+	 */
+	ResourceOwner save = CurrentResourceOwner;
+	StartTransactionCommand();
+	GetTransactionSnapshot();
+	elog(LOG, "probeUpdateConfig called for %d changes", changeCount);
+
+	histrel = heap_open(GpConfigHistoryRelationId,
+						RowExclusiveLock);
+	configrel = heap_open(GpSegmentConfigRelationId,
+						  RowExclusiveLock);
+
+	for (i = 0; i < changeCount; i++)
+	{
+		FtsSegmentStatusChange *change = &changes[i];
+		valid   = (changes[i].newStatus & FTS_STATUS_ALIVE);
+		primary = (changes[i].newStatus & FTS_STATUS_PRIMARY);
+		changelogging = (changes[i].newStatus & FTS_STATUS_CHANGELOGGING);
+
+		if (changelogging)
+		{
+			Assert(primary && valid);
+		}
+
+		Assert((valid || !primary) && "Primary cannot be down");
+
+		/*
+		 * Insert new tuple into gp_configuration_history catalog.
+		 */
+		histvals[Anum_gp_configuration_history_time-1] =
+				TimestampTzGetDatum(GetCurrentTimestamp());
+		histvals[Anum_gp_configuration_history_dbid-1] =
+				Int16GetDatum(changes[i].dbid);
+		snprintf(desc, sizeof(desc),
+				 "FTS: content %d fault marking status %s%s role %c",
+				 change->segindex, valid ? "UP" : "DOWN",
+				 (changelogging) ? " mode: change-tracking" : "",
+				 primary ? GP_SEGMENT_CONFIGURATION_ROLE_PRIMARY : GP_SEGMENT_CONFIGURATION_ROLE_MIRROR);
+		histvals[Anum_gp_configuration_history_desc-1] =
+					CStringGetTextDatum(desc);
+
+		histtuple = heap_form_tuple(RelationGetDescr(histrel), histvals, histnulls);
+		simple_heap_insert(histrel, histtuple);
+		CatalogUpdateIndexes(histrel, histtuple);
+
+		/*
+		 * Find and update gp_segment_configuration tuple.
+		 */
+		ScanKeyInit(&scankey,
+					Anum_gp_segment_configuration_dbid,
+					BTEqualStrategyNumber, F_INT2EQ,
+					Int16GetDatum(changes[i].dbid));
+		sscan = systable_beginscan(configrel, GpSegmentConfigDbidIndexId,
+								   true, SnapshotNow, 1, &scankey);
+		configtuple = systable_getnext(sscan);
+		if (!HeapTupleIsValid(configtuple))
+		{
+			elog(ERROR, "FTS cannot find dbid=%d in %s", changes[i].dbid,
+				 RelationGetRelationName(configrel));
+		}
+		configvals[Anum_gp_segment_configuration_role-1] =
+				CharGetDatum(primary ? GP_SEGMENT_CONFIGURATION_ROLE_PRIMARY : GP_SEGMENT_CONFIGURATION_ROLE_MIRROR);
+		repls[Anum_gp_segment_configuration_role-1] = true;
+		configvals[Anum_gp_segment_configuration_status-1] =
+				CharGetDatum(valid ? GP_SEGMENT_CONFIGURATION_STATUS_UP : GP_SEGMENT_CONFIGURATION_STATUS_DOWN);
+		repls[Anum_gp_segment_configuration_status-1] = true;
+		if (changelogging)
+		{
+			configvals[Anum_gp_segment_configuration_mode-1] =
+					CharGetDatum(GP_SEGMENT_CONFIGURATION_MODE_CHANGETRACKING);
+		}
+		repls[Anum_gp_segment_configuration_mode-1] = changelogging;
+
+		newtuple = heap_modify_tuple(configtuple, RelationGetDescr(configrel),
+									 configvals, confignulls, repls);
+		simple_heap_update(configrel, &configtuple->t_self, newtuple);
+		CatalogUpdateIndexes(configrel, newtuple);
+
+		systable_endscan(sscan);
+		pfree(newtuple);
+		/*
+		 * Update shared memory
+		 */
+		ftsProbeInfo->fts_status[changes[i].dbid] = changes[i].newStatus;
+	}
+	heap_close(histrel, RowExclusiveLock);
+	heap_close(configrel, RowExclusiveLock);
+
+	SIMPLE_FAULT_INJECTOR(FtsWaitForShutdown);
+	/*
+	 * Do not block shutdown.  We will always get a change to update
+	 * gp_segment_configuration in subsequent probes upon database
+	 * restart.
+	 */
+	if (IsFtsShudownRequested())
+	{
+		elog(LOG, "Shutdown in progress, ignoring FTS prober updates.");
+		return false;
+	}
+	CommitTransactionCommand();
+	CurrentResourceOwner = save;
+	return true;
+}
+
+/*
+ * Apply requested segment transitions
+ */
+static void
+updateConfiguration(FtsSegmentStatusChange *changes, int changeEntries)
+{
+	Assert(changes != NULL);
+	Assert(cdb_component_dbs != NULL);
+
+	CdbComponentDatabaseInfo *entryDB = &cdb_component_dbs->entry_db_info[0];
+
+	if (entryDB->dbid != GpIdentity.dbid)
+	{
+		if (gp_log_fts >= GPVARS_VERBOSITY_DEBUG)
+		{
+			elog(LOG, "FTS: advancing to second entry-db.");
+		}
+		entryDB = entryDB + 1;
+	}
+
+	/* if we've gotten a pause or shutdown request, we ignore our probe results. */
+	if (!FtsIsActive())
+	{
+		return;
+	}
+
+	/* update segment configuration */
+	bool commit = probeUpdateConfig(changes, changeEntries);
+
+	if (commit)
+		FtsFailoverFilerep(changes, changeEntries);
+
+	if (gp_log_fts >= GPVARS_VERBOSITY_VERBOSE)
+	{
+		elog(LOG, "FTS: finished segment modifications.");
+	}
+}
+
+#endif
 
 /* EOF */
